+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * It is possible to configure the maximum number of location to fetch in the Stripes config file, typically `stripes.config.js`, using the `maxUnpagedResourceCount` entry in the `config` area. Fixes UIIN-1480.
 * Add visual display when holdings record is suppressed from discovery. Refs UIIN-1378.
+* Add visual display when instance is suppressed from discovery. Refs UIIN-1377 and UIIN-1386.
 
 ## [6.0.0](https://github.com/folio-org/ui-inventory/tree/v6.0.0) (2021-03-18)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.6...v6.0.0)